requirements.txt
docx==0.2.4
Flask==3.1.0
numpy==1.24.4
openai==1.59.3
python-dotenv==1.0.1
python_docx==1.1.2
